requirements.txt
pandas
requests
gspread
oauth2client
PyGithub